refactor(skeleton): break contract import cycles
