+++ packages/bijou-tui-app/src/skeleton-contract.ts
@@ -1,17 +1,13 @@
-import type {
-  BijouContext,
-} from '@flyingrobots/bijou';
-import type {
-  FrameOverlayContext,
-} from '@flyingrobots/bijou-tui';
-import type {
-  SkeletonStatusContext,
-  SkeletonTab,
-} from './skeleton-tab-contract.js';
-import type {
-  SkeletonThemeTokens,
-} from './skeleton-theme-contract.js';
-
+/** Public contract aggregator for the TUI application skeleton. */
+export type {
+  SkeletonOverlayOptions,
+  SkeletonPageConfig,
+  SkeletonPageSpec,
+} from './skeleton-page-contract.js';
+export type {
+  SkeletonMsg,
+  SkeletonPageModel,
+} from './skeleton-state-contract.js';
 export type {
   SkeletonLayoutContext,
   SkeletonRenderContext,
@@ -22,44 +18,3 @@ export type {
   CreateTuiAppSkeletonOptions,
   SkeletonThemeTokens,
 } from './skeleton-theme-contract.js';
-
-/** Page model owned by the stock skeleton shell. */
-export interface SkeletonPageModel {
-  readonly ready: true;
-  readonly drawerOpen: boolean;
-  readonly drawerProgress: number;
-  readonly quitConfirmOpen: boolean;
-}
-
-/** Messages used by the skeleton app shell. */
-export type SkeletonMsg =
-  | { type: 'request-quit' }
-  | { type: 'confirm-quit' }
-  | { type: 'cancel-quit' }
-  | { type: 'toggle-drawer' }
-  | { type: 'drawer-progress'; value: number };
-
-/** Per-page configuration derived from the page spec. */
-export interface SkeletonPageConfig {
-  readonly hasDrawer: boolean;
-  readonly drawerPaneId?: string;
-}
-
-/** A tab and its resolved built-in layout kind. */
-export interface SkeletonPageSpec {
-  readonly tab: SkeletonTab;
-  readonly kind: 'drawer' | 'split' | 'empty' | 'custom';
-}
-
-/** Dependencies needed to assemble shell overlays. */
-export interface SkeletonOverlayOptions {
-  readonly frame: FrameOverlayContext<SkeletonPageModel>;
-  readonly tabsById: ReadonlyMap<string, SkeletonTab>;
-  readonly allTabs: readonly SkeletonTab[];
-  readonly pageConfigs: ReadonlyMap<string, SkeletonPageConfig>;
-  readonly title: string;
-  readonly keyLegend: string;
-  readonly statusMessage?: string | ((ctx: SkeletonStatusContext) => string);
-  readonly tokens?: SkeletonThemeTokens;
-  readonly ctx: BijouContext;
-}
